WAHM, mostly condensing functions to fewer files
